requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
python-multipart==0.0.9
python-dotenv==1.0.1
pydantic==2.7.1
gitpython==3.1.43
networkx==3.3
sentence-transformers==3.0.1
chromadb==0.5.0
ollama==0.2.1
groq==0.9.0
httpx==0.27.0
aiofiles==23.2.1
tqdm==4.66.4
rich==13.7.1